Limit generated chord notes to the range of MIDI notes (i.e., 0-127).
Prior to this change, generated MIDI values would wrap around and result
in unexpected very high or low notes when a generated chord note fell
outside the range.

Generated notes, if they fall outside the range 1-127 (C-2 - G8), are moved
within range by adding or subtracting 12 (one octave) from the offending
note value until it is within range. This can cause the voicing to be
altered from what is implied by the CV and slider input, but the result
will remain in key and be as close as possible to the implied voicing.

Other possible solutions that I rejected:
1. Move the root note down or up until the chord voicing is in range.
Rejected because changing the root note could drastically change a user's
progression.
2. Just don't play the chord notes that are out of range. Rejected as the
implemented solution seemed like it would result in a more musical result.